requirements.txt
av==14.4.0
deepdiff==8.5.0
diffusers==0.33.1
draccus==0.10.0
dynamixel_sdk==3.7.31
einops==0.8.1
feetech_servo_sdk==1.0.0
Flask==3.1.1
gymnasium==0.29.1
huggingface_hub==0.32.4
imageio==2.37.0
jax==0.6.2
jsonlines==4.0.0
lcm==1.5.1
matplotlib==3.10.3
numpy==2.3.0
opencv_python==4.11.0.86
opencv_python_headless==4.11.0.86
orbax==0.1.9
packaging==25.0
pandas==2.3.0
Pillow==11.2.1
pyarrow==20.0.0
pynput==1.8.1
pyrealsense2==2.55.1.6486
pyserial==3.5
pytest==8.4.1
pyzmq==27.0.0
Requests==2.32.4
rerun==1.0.30
safetensors==0.5.3
scikit_learn==1.7.0
scipy==1.15.3
termcolor==3.1.0
torch==2.7.1
torchcodec==0.4.0
torchvision==0.22.1
tqdm==4.67.1
transformers==4.52.4
wandb==0.20.1
